Merge in EPACDX/cdx-oaqps-cef from development to master

* commit 'b5420e1a3a3be277b3a6b21430113113ee66ca68':
  CEF-1485 update to master facility naics table
  CEF-1485 resolving master facility naics bugs

+++ cef-web/src/main/java/gov/epa/cef/web/service/impl/MasterFacilityRecordServiceImpl.java
@@ -29,6 +29,7 @@ import org.springframework.stereotype.Service;
 import gov.epa.cef.web.domain.MasterFacilityRecord;
 import gov.epa.cef.web.domain.NaicsCode;
 import gov.epa.cef.web.config.SLTBaseConfig;
+import gov.epa.cef.web.domain.EmissionsReport;
 import gov.epa.cef.web.domain.FacilityNAICSXref;
 import gov.epa.cef.web.domain.FacilitySite;
 import gov.epa.cef.web.domain.MasterFacilityNAICSXref;
@@ -219,24 +220,27 @@ public class MasterFacilityRecordServiceImpl implements MasterFacilityRecordServ
     
     private void masterFacilityNaicsToFacilityNaics(Long mfrId) {
     	
-    	emissionsReportRepo.findInProgressOrReturnedByMasterFacilityId(mfrId)
-    	.forEach(report -> {
-        	SLTBaseConfig sltConfig = sltConfigHelper.getCurrentSLTConfig(report.getProgramSystemCode().getCode());
-        	report.getFacilitySites().forEach(fs -> {
-            	
-		        if (Boolean.TRUE.equals(sltConfig.getFacilityNaicsEnabled())) {
-		        	facilityNaicsXrefRepo.deleteByFacilitySiteId(fs.getId());
-		        	
-		        	report.getMasterFacilityRecord().getMasterFacilityNAICS().forEach(masterFacilityNAICS -> {
-		        		FacilityNAICSXref facilityNAICS;
-		        		facilityNAICS = facilityNaicsMapper.toFacilityNaicsXref(masterFacilityNAICS);
-		                facilityNAICS.setFacilitySite(fs);
-		                fs.getFacilityNAICS().add(facilityNAICS);
-		            });
-		        }
-		        facilitySiteRepo.save(fs);
-            });
-        });
+    	List<EmissionsReport> erList = emissionsReportRepo.findInProgressOrReturnedByMasterFacilityId(mfrId);
+    	
+    	if (!erList.isEmpty()) {
+    		erList.forEach(report -> {
+	        	SLTBaseConfig sltConfig = sltConfigHelper.getCurrentSLTConfig(report.getProgramSystemCode().getCode());
+	        	report.getFacilitySites().forEach(fs -> {
+	            	
+			        if (Boolean.TRUE.equals(sltConfig.getFacilityNaicsEnabled())) {
+			        	facilityNaicsXrefRepo.deleteByFacilitySiteId(fs.getId());
+			        	
+			        	report.getMasterFacilityRecord().getMasterFacilityNAICS().forEach(masterFacilityNAICS -> {
+			        		FacilityNAICSXref facilityNAICS;
+			        		facilityNAICS = facilityNaicsMapper.toFacilityNaicsXref(masterFacilityNAICS);
+			                facilityNAICS.setFacilitySite(fs);
+			                fs.getFacilityNAICS().add(facilityNAICS);
+			            });
+			        }
+			        facilitySiteRepo.save(fs);
+	            });
+	        });
+    	}
     }
     
     /**
